requirements.txt
altair==5.3.0
matplotlib==3.8.4
pandas==2.2.2
Pillow==10.4.0
plotly==5.21.0
seaborn==0.13.2
streamlit==1.36.0
